+++ src/js/__tests__/util-test.js
@@ -5,6 +5,7 @@ import os from 'os';
 import path from 'path';
 import {
   ensureDir,
+  srcPathsFromMavenDependencies,
   srcPathsFromClasspathStrings,
   isWhitespace,
   isWindows,
@@ -107,5 +108,98 @@ describe('isWhitespace', () => {
     expect(isWhitespace('a')).toBe(false);
     expect(isWhitespace('a b')).toBe(false);
     expect(isWhitespace('a\n')).toBe(false);
+  });
+});
+
+describe('mavenCoordinatesToPath', () => {
+  const osHomedir = os.homedir;
+  beforeAll(() => {
+    os.homedir = jest.fn(() => '/Users/foo');
+  });
+
+  afterAll(() => {
+    os.homedir = osHomedir;
+  });
+
+  it('should decode a group/artifact:version encoded dep to its local path ', () => {
+    expect(
+      srcPathsFromMavenDependencies(['org.clojure/clojurescript:1.9.562']),
+    ).toEqual([
+      path.join.apply(null, [
+        ...(isWindows ? ['C:'] : ['/']),
+        'Users',
+        'foo',
+        '.m2',
+        'repository',
+        'org',
+        'clojure',
+        'clojurescript',
+        '1.9.562',
+        'clojurescript-1.9.562.jar',
+      ]),
+    ]);
+
+    expect(
+      srcPathsFromMavenDependencies([
+        'org.clojure/clojurescript:1.9.562,cljsjs/react:15.5.0-0',
+      ]),
+    ).toEqual([
+      path.join.apply(null, [
+        ...(isWindows ? ['C:'] : ['/']),
+        'Users',
+        'foo',
+        '.m2',
+        'repository',
+        'org',
+        'clojure',
+        'clojurescript',
+        '1.9.562',
+        'clojurescript-1.9.562.jar',
+      ]),
+      path.join.apply(null, [
+        ...(isWindows ? ['C:'] : ['/']),
+        'Users',
+        'foo',
+        '.m2',
+        'repository',
+        'cljsjs',
+        'react',
+        '15.5.0-0',
+        'react-15.5.0-0.jar',
+      ]),
+    ]);
+
+    expect(
+      srcPathsFromMavenDependencies(
+        ['org.clojure/clojurescript:1.9.562'],
+        '~/.m3',
+      ),
+    ).toEqual([
+      path.join.apply(null, [
+        ...(isWindows ? ['C:'] : ['/']),
+        'Users',
+        'foo',
+        '.m3',
+        'org',
+        'clojure',
+        'clojurescript',
+        '1.9.562',
+        'clojurescript-1.9.562.jar',
+      ]),
+    ]);
+
+    expect(srcPathsFromMavenDependencies(['rewrite-clj:0.6.0'])).toEqual([
+      path.join.apply(null, [
+        ...(isWindows ? ['C:'] : ['/']),
+        'Users',
+        'foo',
+        '.m2',
+        'repository',
+        'rewrite-clj',
+        'rewrite-clj',
+        '0.6.0',
+        'rewrite-clj-0.6.0.jar',
+      ]),
+    ]);
   });
 });